[feat/fix_code]: Fix AlbumDeleteDialogFragment
