+++ app/src/main/java/com/teampophory/pophory/feature/signup/SignUpDialogFragment.kt
@@ -14,6 +14,7 @@ import android.view.Window
 import android.view.WindowInsets
 import android.view.WindowManager
 import androidx.fragment.app.DialogFragment
+import com.teampophory.pophory.common.context.dialogWidthPercent
 import com.teampophory.pophory.common.view.viewBinding
 import com.teampophory.pophory.databinding.FragmentSignUpDialogBinding
 
@@ -39,46 +40,9 @@ class SignUpDialogFragment : DialogFragment() {
 
     override fun onResume() {
         super.onResume()
-        context?.let {
-            dialogWidthPercent(it, dialog)
-        }
+        context?.dialogWidthPercent(dialog)
         dialog?.window?.run {
             setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
-        }
-    }
-
-    private fun dialogWidthPercent(context: Context, dialog: Dialog?, percent: Double = 0.8) {
-        val deviceSize = getDeviceSize(context)
-        dialog?.window?.run {
-            val params = attributes
-            params.width = (deviceSize[0] * percent).toInt()
-            attributes = params
-        }
-    }
-
-    private fun getDeviceSize(context: Context): IntArray {
-        val windowManager = context.getSystemService(Context.WINDOW_SERVICE) as WindowManager
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-            val windowMetrics = windowManager.currentWindowMetrics
-            val windowInsets = windowMetrics.windowInsets
-
-            val insets = windowInsets.getInsetsIgnoringVisibility(
-                WindowInsets.Type.navigationBars() or WindowInsets.Type.displayCutout()
-            )
-            val insetsWidth = insets.right + insets.left
-            val insetsHeight = insets.top + insets.bottom
-
-            val bounds = windowMetrics.bounds
-
-            return intArrayOf(bounds.width() - insetsWidth, bounds.height() - insetsHeight)
-        } else {
-            val display = windowManager.defaultDisplay
-            val size = Point()
-
-            display?.getSize(size)
-
-            return intArrayOf(size.x, size.y)
         }
     }
 }
