+++ app/src/main/res/layout/fragment_album_delete_dialog.xml
@@ -4,6 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
+    android:background="@drawable/bg_round_30dp_white"
+    android:paddingHorizontal="16dp"
+    android:paddingTop="16dp"
+    android:paddingBottom="24dp"
     tools:context=".feature.album.detail.AlbumDeleteDialogFragment">
 
     <TextView
@@ -49,16 +53,16 @@
         android:layout_height="wrap_content"
         android:layout_marginHorizontal="16dp"
         android:layout_marginTop="24dp"
+        android:layout_marginBottom="24dp"
         android:background="@drawable/bg_round_30dp_gray40"
         android:gravity="center_horizontal"
         android:paddingVertical="14dp"
         android:text="삭제하기"
         android:textAppearance="@style/TextAppearance.Pophory.PopupButton1"
         android:textColor="@color/white"
+        app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        android:layout_marginBottom="24dp"
         app:layout_constraintTop_toBottomOf="@id/tvReturnButton" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
